requirements.txt
streamlit
langchain-anthropic
python-dotenv
watchdog